Start chromecast SDK only once
